requirements.txt
fastbook
streamlit
streamlit_option_menu
pillow
Keras==2.11.0
tensorflow
timm